+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/state/appliers/AuthorizationCreatedApplier.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.engine.processing.identity;
+package io.camunda.zeebe.engine.state.appliers;
 
 import io.camunda.zeebe.engine.state.TypedEventApplier;
 import io.camunda.zeebe.engine.state.mutable.MutableAuthorizationState;
@@ -13,7 +13,7 @@ import io.camunda.zeebe.engine.state.mutable.MutableProcessingState;
 import io.camunda.zeebe.protocol.impl.record.value.authorization.AuthorizationRecord;
 import io.camunda.zeebe.protocol.record.intent.AuthorizationIntent;
 
-public class AuthorizationCreatedApplier
+public final class AuthorizationCreatedApplier
     implements TypedEventApplier<AuthorizationIntent, AuthorizationRecord> {
   private final MutableAuthorizationState authorizationState;
